Crate cargo_platform

Platform definition used by Cargo.

This defines a Platform type which is used in Cargo to specify a target platform. There are two kinds, a named target like x86_64-apple-darwin, and a “cfg expression” like cfg(any(target_os = "macos", target_os = "ios")).

See examples/matches.rs for an example of how to match against a Platform.

This crate is maintained by the Cargo team for use by the wider ecosystem. This crate follows semver compatibility for its APIs.

Structs

ParseError

Enums

Cfg

A cfg value.

CfgExpr

A cfg expression.

ParseErrorKind

Platform

Platform definition.